% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emojizer.R
\name{emojize_active_file}
\alias{emojize_active_file}
\title{Addin function: Emojize active file.}
\usage{
emojize_active_file()
}
\description{
Emojizes the currently open R code file.
Carefully examine the results after running this function!
}
